Refactored the tek_series_mso features into separate files.
