+++ setup.cfg
@@ -39,27 +39,27 @@ testpaths = test
 
 [options.entry_points]
 visadore.tektronix.mso58 =
-    curve = curvequery._tek_series_mso:TekSeriesCurveFeat
-    default_setup = curvequery._tek_series_mso:TekSeriesDefaultFeat
-    setup = curvequery._tek_series_mso:TekSeriesSetupFeat
-    acquire = curvequery._tek_series_mso:TekSeriesAcquireFeat
+    curve = curvequery._tek_series_mso_curve_feat:TekSeriesCurveFeat
+    default_setup = curvequery._tek_series_mso_setup_feat:TekSeriesDefaultFeat
+    setup = curvequery._tek_series_ms_setup_feat:TekSeriesSetupFeat
+    acquire = curvequery._tek_series_mso_acquire_feat:TekSeriesAcquireFeat
 visadore.tektronix.mso56 =
-    curve = curvequery._tek_series_mso:TekSeriesCurveFeat
-    default_setup = curvequery._tek_series_mso:TekSeriesDefaultFeat
-    setup = curvequery._tek_series_mso:TekSeriesSetupFeat
-    acquire = curvequery._tek_series_mso:TekSeriesAcquireFeat
+    curve = curvequery._tek_series_mso_curve_feat:TekSeriesCurveFeat
+    default_setup = curvequery._tek_series_mso_setup_feat:TekSeriesDefaultFeat
+    setup = curvequery._tek_series_ms_setup_feat:TekSeriesSetupFeat
+    acquire = curvequery._tek_series_mso_acquire_feat:TekSeriesAcquireFeat
 visadore.tektronix.mso54 =
-    curve = curvequery._tek_series_mso:TekSeriesCurveFeat
-    default_setup = curvequery._tek_series_mso:TekSeriesDefaultFeat
-    setup = curvequery._tek_series_mso:TekSeriesSetupFeat
-    acquire = curvequery._tek_series_mso:TekSeriesAcquireFeat
+    curve = curvequery._tek_series_mso_curve_feat:TekSeriesCurveFeat
+    default_setup = curvequery._tek_series_mso_setup_feat:TekSeriesDefaultFeat
+    setup = curvequery._tek_series_ms_setup_feat:TekSeriesSetupFeat
+    acquire = curvequery._tek_series_mso_acquire_feat:TekSeriesAcquireFeat
 visadore.tektronix.mso46 =
-    curve = curvequery._tek_series_mso:TekSeriesCurveFeat
-    default_setup = curvequery._tek_series_mso:TekSeriesDefaultFeat
-    setup = curvequery._tek_series_mso:TekSeriesSetupFeat
-    acquire = curvequery._tek_series_mso:TekSeriesAcquireFeat
+    curve = curvequery._tek_series_mso_curve_feat:TekSeriesCurveFeat
+    default_setup = curvequery._tek_series_mso_setup_feat:TekSeriesDefaultFeat
+    setup = curvequery._tek_series_ms_setup_feat:TekSeriesSetupFeat
+    acquire = curvequery._tek_series_mso_acquire_feat:TekSeriesAcquireFeat
 visadore.tektronix.mso44 =
-    curve = curvequery._tek_series_mso:TekSeriesCurveFeat
-    default_setup = curvequery._tek_series_mso:TekSeriesDefaultFeat
-    setup = curvequery._tek_series_mso:TekSeriesSetupFeat
-    acquire = curvequery._tek_series_mso:TekSeriesAcquireFeat
+    curve = curvequery._tek_series_mso_curve_feat:TekSeriesCurveFeat
+    default_setup = curvequery._tek_series_mso_setup_feat:TekSeriesDefaultFeat
+    setup = curvequery._tek_series_ms_setup_feat:TekSeriesSetupFeat
+    acquire = curvequery._tek_series_mso_acquire_feat:TekSeriesAcquireFeat
